conf_llm_agent_level/pyproject.toml
[project]
name = "conf-llm-agent-level"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hamza Atiq", email = "66232332+Hamza-Atiq@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
conf-llm-agent-level = "conf_llm_agent_level:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
